+++ src/assets/scss/components/index.scss
@@ -21,4 +21,4 @@
 @import 'footer';
 @import 'alerts';
 @import 'search';
-@import 'productSelect.scss';
+@import 'productList.scss';
