+++ src/assets/scss/components/productList.scss
@@ -1,8 +1,9 @@
-.product-select-container {
+.product-list-container {
   text-align: center;
 
-  .product-select {
+  .product-list {
     margin-bottom: 10px;
+    min-height: 160px;
     width: 100%;
   }
 }
@@ -11,10 +12,11 @@
   border-radius: 5px;
   box-shadow: 0px 2px 15px $shadow-color;
   margin: 8px 15px;
+  min-width: 240px;
   padding: 10px;
   text-transform: uppercase;
   transition: 0.2s;
-  width: 260px;
+  width: 28%;
 
   .circle-container {
     width: 100%;
@@ -30,10 +32,11 @@
   }
 
   .product-logo {
-    width: 230px;
+    min-height: 80px;
+    width: 260px;
 
     img {
-      width: 230px;
+      width: 260px;
     }
   }
