+++ src/components/Stories.js
@@ -9,7 +9,7 @@ import SearchInput from '../modules/SearchInput'
 
 import Lists from '../utils/Lists'
 import userStory from '../services/user_story'
-import ProductSelect from './ProductSelect'
+import ProductList from './ProductList'
 
 const Stories = ({ authorId, followerId }) => {
   const [currentStateSelected, selectState] = useState('Under consideration')
@@ -20,13 +20,9 @@ const Stories = ({ authorId, followerId }) => {
 
   const [status, setStatus] = useState('Under consideration')
 
-  const [product, setProduct] = useState('All')
-
   const [sort, setSort] = useState('Most Voted')
 
   const [category, setCategory] = useState('All')
-
-  const [products, setProducts] = useState(null)
 
   const [categories, setCategories] = useState([])
 
@@ -80,7 +76,6 @@ const Stories = ({ authorId, followerId }) => {
     fetchStoryCount()
   }, [
     currentStateSelected,
-    product,
     categoryQuery,
     productQuery,
     searchQuery,
@@ -90,11 +85,6 @@ const Stories = ({ authorId, followerId }) => {
   ])
 
   useEffect(() => {
-    if (product !== 'All') {
-      setProductQuery(`product : {Name: "${product}"}`)
-    } else {
-      setProductQuery(``)
-    }
     if (category !== 'All') {
       setCategoryQuery(`Category : "${category}"`)
     } else {
@@ -106,7 +96,7 @@ const Stories = ({ authorId, followerId }) => {
     if (userTerm === '') {
       setAuthorQuery('')
     }
-  }, [product, category, searchTerm, userTerm])
+  }, [category, searchTerm, userTerm])
 
   useEffect(() => {
     const fetchStories = async () => {
@@ -133,16 +123,6 @@ const Stories = ({ authorId, followerId }) => {
     authorId,
     followerId
   ])
-
-  useEffect(() => {
-    const fetchProducts = async () => {
-      const response = await userStory.getProducts()
-      return response.data.data.product !== null
-        ? setProducts([...response.data.data.products])
-        : setProducts([])
-    }
-    fetchProducts()
-  }, [])
 
   useEffect(() => {
     const fetchCategories = async () => {
@@ -180,11 +160,7 @@ const Stories = ({ authorId, followerId }) => {
 
   return (
     <div>
-      <ProductSelect
-        product={product}
-        products={products}
-        setProduct={setProduct}
-      />
+      <ProductList setProductQuery={setProductQuery} />
       <div className='roadmap-container'>
         <div className='roadmap'>
           {Lists.stateList &&
@@ -251,18 +227,13 @@ const Stories = ({ authorId, followerId }) => {
         </div>
       </div>
       <div className='stories-div'>
-        <StoriesList
-          stories={stories}
-          state={currentStateSelected}
-          product={product}
-          isLoading={promiseInProgress}
-        />
+        <StoriesList stories={stories} isLoading={promiseInProgress} />
       </div>
       <Pagination
         getPage={getPage}
         storyCount={storyCount}
         status={currentStateSelected}
-        product={product}
+        productQuery={productQuery}
       />
     </div>
   )
