+++ src/services/user_story.js
@@ -198,6 +198,18 @@ const userStory = {
     }
     return apiCall('/graphql', productQuery)
   },
+  getProductCount: () => {
+    const productCountQuery = {
+      query: `query {
+        productsConnection {
+          aggregate {
+            count
+          }
+        }
+      }`
+    }
+    return apiCall('/graphql', productCountQuery)
+  },
   getProductsWithTemplates: () => {
     const productQuery = {
       query: `query {
